Add bootstrap views, and moderator post flow.
Moderator can accept/decline or add suggestions for a post.
He has a seperate view for it where all posts will be listed.
View is basic bootstrap with some nav bar and sidebar (not done 100% 
yet).
Added feature tests for moderator flow.
Added callback for authentication for now as the fastest way to do it,
later on this will get changed to tyrant I gues?